Fix bugs when camera unmounts before starting
If the camera was unmounted before it had finished starting, there were two
issues. First, a setState call was made which put an error in the console about
setting state on unmounted components. Second, the media stream was not stopped,
so the user's camera would stay on. This commit fixes that problem.

Fixes #127.